Switched the int ids to the long type

+++ AuthService/Repositories/User/IUserRepository.cs
@@ -6,7 +6,7 @@ namespace BackGazprom.Repositories;
 public interface IUserRepository
 {
     public IQueryable<User> GetUsers();
-    public Task<User?> GetUserById(int id);
+    public Task<User?> GetUserById(long id);
     public Task<User?> GetUserByUsername(string username);
     public Task<User?> GetUserByEmail(string email);
     public Task<bool> CreateUser(User user);
